memory-engine/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
openai==1.3.0
redis==5.0.0
chromadb==0.4.18
python-multipart==0.0.6
httpx==0.25.0
python-dotenv==1.0.0
tenacity==8.2.3
